파일 이름들 수정과 메인 부분 화면들 연결
