+++ app/src/main/java/com/cookandroid/medication_helper/AgeForbiddenListActivity.java
@@ -7,12 +7,12 @@ import android.widget.Button;
 
 import androidx.appcompat.app.AppCompatActivity;
 
-public class medicinelist extends AppCompatActivity {
+public class AgeForbiddenListActivity extends AppCompatActivity {
 
     @Override
     public void onBackPressed() {
         super.onBackPressed();
-        Intent Back = new Intent(medicinelist.this, medicCheck.class);
+        Intent Back = new Intent(AgeForbiddenListActivity.this, MedicCheckActivity.class);
         Back.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
         startActivity(Back);
         finish();
@@ -21,19 +21,20 @@ public class medicinelist extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.medicinelist);
+        setContentView(R.layout.activity_ageforbiddenlist);
         setTitle("Medication Helper");
 
-        Button btnBack = findViewById(R.id.btnback1);
+        Button btnBack = findViewById(R.id.btnback3);
 
         btnBack.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View view) {
-                Intent intent = new Intent(medicinelist.this, medicCheck.class);
+                Intent intent = new Intent(AgeForbiddenListActivity.this, MedicCheckActivity.class);
                 startActivity(intent);
                 finish();
             }
         });
 
     }
+
 }
