+++ app/src/main/java/com/cookandroid/medication_helper/ComForbiddenListActivity.java
@@ -1,0 +1,26 @@
+package com.cookandroid.medication_helper;
+
+import android.content.Intent;
+import android.os.Bundle;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class ComForbiddenListActivity extends AppCompatActivity {
+
+    @Override
+    public void onBackPressed() {
+        super.onBackPressed();
+        Intent Back = new Intent(ComForbiddenListActivity.this, MedicCheckActivity.class);
+        Back.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+        startActivity(Back);
+        finish();
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_comforbiddenlist);
+        setTitle("Medication Helper");
+
+    }
+}
