+++ app/src/main/java/com/cookandroid/medication_helper/ListViewAdapterData.java
@@ -1,6 +1,6 @@
 package com.cookandroid.medication_helper;
 
-public class listViewAdapterData {
+public class ListViewAdapterData {
 
     private String uName;
     private String uBirth;
