+++ app/src/main/java/com/cookandroid/medication_helper/MedicateFrequencyActivity.java
@@ -1,0 +1,4 @@
+package com.cookandroid.medication_helper;
+
+public class MedicateFrequencyActivity {
+}
