+++ app/src/main/java/com/cookandroid/medication_helper/TimeRegisterActivity.java
@@ -1,0 +1,28 @@
+package com.cookandroid.medication_helper;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+public class TimeRegisterActivity extends AppCompatActivity {
+    @Override
+    public void onBackPressed() {
+        super.onBackPressed();
+        Intent BackToMain = new Intent(TimeRegisterActivity.this, MainActivity.class);
+        BackToMain.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+        startActivity(BackToMain);
+        finish();
+    }
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_timeregister);
+        setTitle("Medication Helper");
+
+
+    }
+}
