+++ app/src/main/res/drawable/green_cross_size.xml
