+++ app/src/main/res/layout/activity_comforbiddenlist.xml
@@ -1,0 +1,52 @@
+<?xml version="1.0" encoding="utf-8"?>
+<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical">
+
+    <LinearLayout
+        android:layout_width="fill_parent"
+        android:layout_height="fill_parent"
+        android:orientation="vertical"
+        >
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:text="병용 금기 약물"
+            android:textColor="#036635"
+            android:textSize="30sp"
+            android:layout_gravity="center_horizontal|top"
+            android:layout_margin="10dp" />
+
+        <ListView
+            android:id="@+id/combinationXList"
+            android:layout_width="match_parent"
+            android:layout_height="500dp"
+            android:background="@drawable/green_border"
+            android:layout_marginBottom="5dp"/>
+
+        <TextView
+            android:text=" 금기 성분"
+            android:textColor="#036635"
+            android:textSize="30dp"
+            android:layout_width="match_parent"
+            android:layout_height="35dp"
+            android:layout_marginBottom="5dp"
+            />
+        <TextView
+            android:id="@+id/fobiddenIngredeint"
+            android:layout_width="match_parent"
+            android:layout_height="60dp"
+            android:textSize="50dp"
+            android:background="@drawable/green_border"/>
+
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="70dp"
+            android:layout_gravity="center">
+
+            //여기 버튼이 있었던거 같은데 제가 없어 버렸네여 나중에 추가하도록
+            //하겠습니다
+        </RelativeLayout>
+    </LinearLayout>
+</ScrollView>
