+++ app/src/main/res/layout/activity_login.xml
@@ -1,0 +1,84 @@
+<?xml version="1.0" encoding="utf-8"?>
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".LoginActivity">
+
+    <LinearLayout
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        android:layout_marginTop="200dp"
+        android:orientation="vertical">
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_marginBottom="20dp"
+            android:text="회원 로그인"
+            android:textColor="#036635"
+            android:textSize="30sp"
+            android:layout_gravity="center_horizontal|top" />
+        <EditText
+            android:layout_width="wrap_content"
+            android:layout_height="50dp"
+            android:layout_marginBottom="10dp"
+            android:id="@+id/editID"
+            android:layout_gravity="center"
+            android:ems="12"
+            android:hint="  아이디"
+            android:inputType="textPersonName"
+            android:text=""
+            android:background="@drawable/green_border"/>
+
+        <EditText
+            android:layout_width="wrap_content"
+            android:layout_height="50dp"
+            android:id="@+id/editPW"
+            android:layout_gravity="center"
+            android:ems="12"
+            android:hint="  비밀번호"
+            android:inputType="textPassword"
+            android:background="@drawable/green_border"/>
+
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/btnlogin"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_horizontal"
+            android:layout_marginLeft="130dp"
+            android:layout_marginTop="20dp"
+            android:layout_marginRight="130dp"
+            android:background="@drawable/green_border"
+            android:text="로그인"
+            android:textColor="#036635"
+            android:textSize="17sp" />
+
+        <androidx.appcompat.widget.AppCompatButton
+            android:id="@+id/btnsignin"
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center_horizontal"
+            android:layout_marginLeft="130dp"
+            android:layout_marginTop="10dp"
+            android:layout_marginRight="130dp"
+            android:layout_marginBottom="150dp"
+            android:background="@drawable/green_border"
+            android:text="회원가입"
+            android:textColor="#036635"
+            android:textSize="17sp" />
+
+        <TextView
+            android:layout_width="wrap_content"
+            android:layout_height="wrap_content"
+            android:layout_gravity="center|bottom"
+            android:drawableStart="@drawable/green_cross_size"
+            android:text="3분카레 약국"
+            android:textColor="#036635"
+            android:textSize="30sp"
+            android:drawablePadding="5dp"
+            android:layout_margin="10dp"/>
+
+    </LinearLayout>
+</FrameLayout>
