+++ app/src/main/res/layout/activity_main.xml
@@ -11,7 +11,22 @@
         android:id="@+id/content_layout"
         android:layout_width="match_parent"
         android:layout_height="0dp"
-        android:layout_weight="1" />
+        android:layout_weight="1" >
+
+        <LinearLayout
+            android:layout_width="match_parent"
+            android:layout_height="wrap_content"
+            android:orientation="vertical"
+            android:id="@+id/main">
+        </LinearLayout>
+
+        <RelativeLayout
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            android:id="@+id/fragment">
+
+        </RelativeLayout>
+    </FrameLayout>
 
     <com.google.android.material.bottomnavigation.BottomNavigationView
         android:id="@+id/bottomNav"
@@ -23,5 +38,5 @@
         android:layout_centerHorizontal="true"
         app:itemIconTint="@drawable/navcolor"
         app:itemIconSize="30dp"
-        app:menu="@menu/mainnav" />
+        app:menu="@menu/nav" />
 </LinearLayout>
