+++ app/src/main/res/layout/activity_pregnantforbiddenlist.xml
@@ -12,7 +12,7 @@
         <TextView
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:text="병용 금기 약물"
+            android:text="임산부 금기 약물"
             android:textColor="#036635"
             android:textSize="30sp"
             android:layout_gravity="center_horizontal|top"
@@ -22,9 +22,8 @@
             android:id="@+id/combinationXList"
             android:layout_width="match_parent"
             android:layout_height="500dp"
-            android:background="@drawable/green_box"
+            android:background="@drawable/green_border"
             android:layout_marginBottom="5dp"/>
-
         <TextView
             android:text=" 금기 성분"
             android:textColor="#036635"
@@ -38,7 +37,7 @@
             android:layout_width="match_parent"
             android:layout_height="60dp"
             android:textSize="50dp"
-            android:background="@drawable/green_box"/>
+            android:background="@drawable/green_border"/>
 
         <RelativeLayout
             android:layout_width="match_parent"
@@ -46,13 +45,13 @@
             android:layout_gravity="center">
 
             <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btnback2"
+                android:id="@+id/btnback4"
                 android:layout_width="120dp"
                 android:layout_height="50dp"
                 android:layout_gravity="center"
                 android:layout_marginLeft="220dp"
                 android:layout_marginTop="10dp"
-                android:background="@drawable/green_box"
+                android:background="@drawable/green_border"
                 android:text="뒤로 가기"
                 android:textColor="#036635"
                 android:textSize="20dp" />
