+++ app/src/main/res/layout/fragment_home.xml
@@ -1,10 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="fill_parent"
-    android:layout_height="fill_parent"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:orientation="vertical"
-    android:theme="@style/Theme.AppCompat.Light.NoActionBar">
+    android:theme="@style/Theme.AppCompat.Light.NoActionBar"
+    tools:context=".HomeFragment">
 
     <FrameLayout
         android:layout_width="match_parent"
@@ -15,7 +17,7 @@
             android:layout_height="wrap_content"
             android:layout_gravity="center_horizontal"
             android:layout_margin="10dp"
-            android:drawableStart="@drawable/drawablesize"
+            android:drawableStart="@drawable/green_cross_size"
             android:drawablePadding="3dp"
             android:text="3분카레 약국"
             android:textColor="#036635"
@@ -29,7 +31,7 @@
             <TextView
                 android:layout_width="match_parent"
                 android:layout_height="match_parent"
-                android:background="@drawable/green_box"
+                android:background="@drawable/green_border"
                 android:padding="10dp"
                 android:text="복약 정보"
                 android:textColor="#036635"
@@ -43,7 +45,7 @@
                 android:layout_marginLeft="80dp"
                 android:layout_marginTop="200px"
                 android:layout_marginRight="80dp"
-                android:background="@drawable/green_box"
+                android:background="@drawable/green_border"
                 android:text="복약 정보 등록"
                 android:textColor="#036635"
                 android:textSize="17sp" />
@@ -56,7 +58,7 @@
                 android:layout_marginLeft="80dp"
                 android:layout_marginTop="400px"
                 android:layout_marginRight="80dp"
-                android:background="@drawable/green_box"
+                android:background="@drawable/green_border"
                 android:text="복약 정보 조회"
                 android:textColor="#036635"
                 android:textSize="17sp" />
@@ -70,10 +72,10 @@
                 android:layout_marginTop="225dp"
                 android:layout_marginRight="80dp"
                 android:layout_marginBottom="70px"
-                android:background="@drawable/green_box"
+                android:background="@drawable/green_border"
                 android:text="복약 알람 설정"
                 android:textColor="#036635"
                 android:textSize="17sp" />
         </FrameLayout>
     </FrameLayout>
-</RelativeLayout>
+</FrameLayout>
