--- app/src/main/res/layout/medicinelist.xml
@@ -1,61 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <LinearLayout
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:orientation="vertical"
-        >
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="처방약 목록"
-            android:textColor="#036635"
-            android:textSize="30sp"
-            android:layout_gravity="center_horizontal|top"
-            android:layout_margin="10dp" />
-
-        <ListView
-            android:id="@+id/medicationlist"
-            android:layout_width="match_parent"
-            android:layout_height="600dp"
-            android:background="@drawable/green_box"
-            android:layout_marginBottom="5dp"/>
-
-
-        <RelativeLayout
-            android:layout_width="match_parent"
-            android:layout_height="70dp"
-            android:layout_gravity="center">
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btndelete"
-                android:layout_width="120dp"
-                android:layout_height="50dp"
-                android:layout_gravity="center"
-                android:layout_marginLeft="80dp"
-                android:layout_marginTop="10dp"
-                android:background="@drawable/green_box"
-                android:text="목록 삭제"
-                android:textColor="#036635"
-                android:textSize="20dp" />
-
-            <androidx.appcompat.widget.AppCompatButton
-                android:id="@+id/btnback1"
-                android:layout_width="120dp"
-                android:layout_height="50dp"
-                android:layout_gravity="center"
-                android:layout_marginLeft="220dp"
-                android:layout_marginTop="10dp"
-                android:background="@drawable/green_box"
-                android:text="뒤로 가기"
-                android:textColor="#036635"
-                android:textSize="20dp" />
-
-        </RelativeLayout>
-    </LinearLayout>
-
-</ScrollView>
