--- app/src/main/res/layout/userregister.xml
@@ -1,178 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity"
-    android:orientation="vertical">
-
-    <LinearLayout
-        android:layout_width="fill_parent"
-        android:layout_height="fill_parent"
-        android:orientation="vertical" >
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="25dp"
-            android:text="회원 가입"
-            android:textSize="40dp"
-            android:textStyle="bold"
-            android:textColor="#036635"
-            android:drawableStart="@drawable/drawablesize"
-            android:drawableEnd="@drawable/drawablesize">
-        </TextView>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="25dp"
-            android:text="ID (20자 이내)"
-            android:textSize="30dp"
-            android:textStyle="bold"
-            android:textColor="#036635">
-        </TextView>
-
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/E_ID"
-            android:layout_gravity="center"
-            android:gravity="center"
-            android:layout_marginTop="15dp"
-            android:ems = "7">
-        </EditText>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center"
-            android:layout_marginTop="25dp"
-            android:text="비밀번호"
-            android:textSize="30dp"
-            android:textStyle="bold"
-            android:textColor="#036635">
-        </TextView>
-
-        <EditText
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/E_Pass"
-            android:layout_gravity="center"
-            android:gravity="center"
-            android:layout_marginTop="15dp"
-            android:ems = "7">
-        </EditText>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text = "이름"
-            android:layout_gravity="center"
-            android:layout_marginTop="25dp"
-            android:textSize="30dp"
-            android:textStyle="bold"
-            android:textColor="#036635">
-        </TextView>
-
-        <EditText
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/E_name"
-            android:layout_gravity="center"
-            android:gravity="center"
-            android:layout_marginTop="15dp"
-            android:ems = "7">
-        </EditText>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text = "생년월일"
-            android:layout_gravity="center"
-            android:layout_marginTop="25dp"
-            android:textSize="30dp"
-            android:textStyle="bold"
-            android:textColor="#036635">
-
-        </TextView>
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/BtnBirthChoose"
-            android:text="선택"
-            android:layout_gravity="center"
-            android:textSize="30dp"
-            android:layout_marginTop="30dp"
-            android:layout_marginLeft="130dp"
-            android:layout_marginRight="130dp"
-            android:background="@drawable/green_box"
-            android:backgroundTint="@android:color/holo_green_light"/>
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text = "성별"
-            android:layout_gravity="center"
-            android:layout_marginTop="30dp"
-            android:textSize="30dp"
-            android:textStyle="bold">
-        </TextView>
-
-        <RadioGroup
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/RG"
-            android:gravity="center"
-            android:layout_marginTop="15dp"
-            android:orientation="horizontal">
-
-            <RadioButton
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:id="@+id/RB_man"
-                android:text="남"
-                android:textSize="30dp"
-                android:layout_marginRight="80dp">
-            </RadioButton>
-            <RadioButton
-                android:layout_width="wrap_content"
-                android:layout_height="wrap_content"
-                android:text="여"
-                android:textSize="25dp"
-                android:id="@+id/RB_woman">
-            </RadioButton>
-
-        </RadioGroup>
-
-        <androidx.appcompat.widget.AppCompatButton
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:id="@+id/BtnComplete"
-            android:layout_marginTop="30dp"
-            android:layout_marginLeft="130dp"
-            android:layout_marginRight="130dp"
-            android:layout_gravity="center"
-            android:text="설정 완료"
-            android:textSize="30dp"
-            android:background="@drawable/green_box"
-            android:backgroundTint="@android:color/holo_green_light">
-        </androidx.appcompat.widget.AppCompatButton>
-
-
-        <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_gravity="center|bottom"
-            android:drawableStart="@drawable/drawablesize"
-            android:text="3분카레 약국"
-            android:textColor="#036635"
-            android:textSize="30sp"
-            android:drawablePadding="5dp"
-            android:layout_margin="10dp"/>
-    </LinearLayout>
-
-</ScrollView>
